Cleaned main.js
