+++ src/App.vue
@@ -34,11 +34,14 @@
 </template>
 
 <script>
+import httpClient from '@/services/http-client';
+
 import AppHeader from '@/components/Header';
 import Export from '@/views/Export';
 import Fulltext from '@/views/Fulltext';
 import Help from '@/views/Help';
 import Info from '@/views/Info';
+
 import Scan from '@/views/Scan';
 import Thumbnails from '@/views/Thumbnails';
 import Toc from '@/views/Toc';
@@ -61,6 +64,68 @@ export default {
 		hasToc() {
 			return !!(this.$root.manifest.structures && this.$root.manifest.structures.length);
 		},
+	},
+	mounted() {
+		httpClient.interceptors.request.use((request) => {
+			this.$root.loading += 1;
+			return request;
+		});
+
+		httpClient.interceptors.response.use((response) => {
+			if (this.$root.loading > 0) this.$root.loading -= 1;
+			return response;
+		}, (error) => {
+			this.$root.loading = 0;
+			return Promise.reject(error);
+		});
+
+		// Manifest URL in tifyOptions trumps query param
+		// Support `manifestUrl` for backward compatibility if `manifest` is not set
+		this.$root.manifestUrl = this.$root.options.manifest
+				|| this.$root.getQueryParam('manifest')
+				|| this.$root.getQueryParam('manifestUrl');
+
+		if (!this.$root.manifestUrl) {
+			this.$root.error = 'Missing query parameter or option: manifest';
+			return;
+		}
+
+		if (this.$root.options.manifest && this.$root.params.manifest) {
+			this.$root.error = 'Setting manifest via query parameter is disabled';
+		}
+
+		// Load manifest
+		httpClient.get(this.$root.manifestUrl).then((response) => {
+			this.$root.manifest = response.data;
+
+			// Merging user-set query params with defaults
+			this.$root.params = this.$root.getParams();
+
+			window.addEventListener('popstate', () => {
+				this.$root.params = this.$root.getParams();
+			});
+
+			if (this.$root.options.title) {
+				window.document.title = `${
+					this.$root.convertValueToArray(this.$root.manifest.label)[0]} | ${this.$root.options.title
+				}`;
+			}
+		}, (error) => {
+			const status = (error.response ? error.response.statusText : error.message);
+			this.$root.error = `Error loading IIIF manifest: ${status}`;
+		});
+
+		// Load translation
+		if (this.$root.options.language !== 'en') {
+			const translationUrl = `/translations/${this.$root.options.language}.json`;
+			httpClient.get(translationUrl).then((response) => {
+				this.messages = response.data;
+			}, (error) => {
+				const status = (error.response ? error.response.statusText : error.message);
+				this.$root.error = `Error loading translation ${this.$root.options.language}: ${status}`;
+				console.warn(this.$root.error); // eslint-disable-line no-console
+			});
+		}
 	},
 };
 </script>
