+++ src/views/Info.vue
@@ -4,7 +4,7 @@
 
 		<div v-if="manifest.label" class="tify-info_section -title">
 			<h3 class="tify-info_heading">{{ 'Title'|trans }}</h3>
-			<div :key="label" v-for="label in $root.iiifConvertToArray(manifest.label)">
+			<div :key="label" v-for="label in getLabels(manifest.label)">
 				{{ label }}
 			</div>
 		</div>
@@ -32,7 +32,7 @@
 			<h3>{{ 'Description'|trans }}</h3>
 			<div
 				:key="index"
-				v-for="(description, index) in $root.iiifConvertToArray(manifest.description)"
+				v-for="(description, index) in getLabels(manifest.description)"
 				v-html="description"/>
 		</div>
 
@@ -72,7 +72,7 @@
 
 		<div v-if="manifest.attribution" class="tify-info_section -attribution">
 			<h3>{{ 'Provided by'|trans }}</h3>
-			<div :key="index" v-for="(item, index) in $root.iiifConvertToArray(manifest.attribution)" v-html="item"/>
+			<div :key="index" v-for="(item, index) in getLabels(manifest.attribution)" v-html="item"/>
 		</div>
 
 		<div v-if="manifest.logo" class="tify-info_section -logo">
@@ -89,7 +89,6 @@
 
 <script>
 // TODO: Handle and display manifest.service, see http://iiif.io/api/presentation/2.1/#service
-
 import MetadataList from '@/components/MetadataList';
 
 import structures from '@/mixins/structures';
@@ -108,7 +107,7 @@ export default {
 	},
 	computed: {
 		license() {
-			return this.manifest.license ? this.$root.iiifConvertToArray(this.manifest.license) : [];
+			return this.manifest.license ? this.getLabels(this.manifest.license) : [];
 		},
 		logoId() {
 			return (this.manifest.logo['@id'] ? this.manifest.logo['@id'] : this.manifest.logo);
@@ -117,7 +116,7 @@ export default {
 			return this.$root.manifest;
 		},
 		related() {
-			return this.manifest.related ? this.$root.iiifConvertToArray(this.manifest.related) : [];
+			return this.manifest.related ? this.getLabels(this.manifest.related) : [];
 		},
 	},
 	methods: {
@@ -127,6 +126,9 @@ export default {
 		isUrl(string) {
 			// Poor man's URL check
 			return /^https?:\/\//.test(string);
+		},
+		getLabels(value) {
+			return this.$root.convertValueToArray(value);
 		},
 	},
 	watch: {
